refactor: change package name to org.rucca.cheese
