+++ src/main/kotlin/org/rucca/cheese/common/InternalServerError.kt
@@ -1,9 +1,9 @@
-package cn.edu.ruc.cheese.backend.common
+package org.rucca.cheese.common
 
 import org.springframework.http.HttpStatus
 
 class InternalServerError :
-        BaseError(
+        org.rucca.cheese.common.BaseError(
                 HttpStatus.INTERNAL_SERVER_ERROR,
                 "InternalServerError",
                 "The server encountered an unexpected error. Please try again later.")
