+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=backend
+spring.application.name=cheese
 
 # Default database configuration
 # If you follow README.md and use docker to setup the database, this configuration should work
